Added Multiple Windows
- Added Calendar and Main Activity
- Added a debug button to go to Calendar
- Added a Calendar object to Calendar
- Added a "Task" button to go to Tasks Activity
- Connect all three activities via button

** CANNOT DEBUG DUE TO HARDWARE **
