+++ src/Scheducation/app/src/main/AndroidManifest.xml
@@ -11,7 +11,13 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".MainActivity">
+        <activity android:name=".TaskActivity" android:label="@string/app_name">
+
+        </activity>
+        <activity android:name=".CalendarActivity" android:label="@string/app_name">
+
+        </activity>
+        <activity android:name=".MainActivity" android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
