+++ src/Scheducation/app/src/main/res/layout/activity_main.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <Button
+        android:id="@+id/buttonCalendar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_vertical"
+        android:text="@string/go_to_calendar"
+        android:translationX="25sp"
+        android:translationY="25sp"
+        android:onClick="gotoCalendar"/>
+</LinearLayout>
